Cambio de usar Bytes a usar enteros
Esto con el fin de ampliar la cantidad de bytes que podemos usar

+++ src/logica/SistemaSAP.java
@@ -44,7 +44,7 @@ public class SistemaSAP implements IClockObserver {
     private IRegistro registroMAR;
     private IRegistro bus;
     private ProgramCounter programCounter;
-    private byte stepCount;
+    private int stepCount;
     private Memoria RAM;
     private EventLog log;
     private ALU alu;
@@ -123,7 +123,7 @@ public class SistemaSAP implements IClockObserver {
         }
 
         // Nada está poniendo su valor en el bus, así que se limpia
-        this.bus.setValor((byte) 0);
+        this.bus.setValor(0);
 
         // Informamos a la vista que vuelva a pintar el bus
         this.notificarCambioBUS();
@@ -134,19 +134,19 @@ public class SistemaSAP implements IClockObserver {
 
     private TipoInstruccion decodificarIR() {
         // Obtener el valor almacenado en el registro de instrucciones
-        byte instruccion = this.registroIR.getValor();
+        int instruccion = this.registroIR.getValor();
 
         // Descartar los cuatro bits menos significativos
-        instruccion = (byte) (instruccion & 0b11110000);
+        instruccion = (int) (instruccion & 0b11110000);
 
         // Analizar el valor
         return decodificarInstruccion(instruccion);
     }
 
-    // Método auxiliar que analiza un byte y encuentra su tipo de instrucción
-    public TipoInstruccion decodificarInstruccion(byte i) {
-        byte instruccion = (byte) (i>>4);
-        instruccion = (byte) (instruccion & 0b00001111);
+    // Método auxiliar que analiza un int y encuentra su tipo de instrucción
+    public TipoInstruccion decodificarInstruccion(int i) {
+        int instruccion = (int) (i>>4);
+        instruccion = (int) (instruccion & 0b00001111);
         //System.out.printf("INS: %x\n", instruccion);
         switch (instruccion) {
             case 0:
@@ -178,12 +178,12 @@ public class SistemaSAP implements IClockObserver {
          
     }
 
-    public void analizarInstruccion(byte address) {
+    public void analizarInstruccion(int address) {
         // Comience a construir una entrada para el registro
         String log = "[" + address + "]\t";
 
         // Primero, agregue la instrucción al Log
-        byte instructionVal = (byte) (this.getRAM().getData()[address] & 0b11110000);
+        int instructionVal = (int) (this.getRAM().getData()[address] & 0b11110000);
         TipoInstruccion t = decodificarInstruccion(instructionVal);
 
         // Manejar el resultado de la instrucción decodificada
@@ -526,13 +526,13 @@ public class SistemaSAP implements IClockObserver {
                 this.notificarCambioBUS();
             }
             if (this.lineasControl[RO]) {
-                this.bus.setValor((byte) this.RAM.memoryOut());
+                this.bus.setValor((int) this.RAM.memoryOut());
                 EventLog.getEventLog().addEntrada("Valor de RAM en el bus");
                 this.notificarCambioBUS();
             }
             if (this.lineasControl[IO]) {
                 // Coloque 4 bits menos significativos del registro de instrucciones en el bus
-                this.bus.setValor((byte) (0b00001111 & this.registroIR.getValor()));
+                this.bus.setValor((int) (0b00001111 & this.registroIR.getValor()));
                 this.notificarCambioBUS();
                 EventLog.getEventLog().addEntrada("Valor del Instruction Register en el bus (4 Bits)");
             }
@@ -602,7 +602,7 @@ public class SistemaSAP implements IClockObserver {
 
             }
             if (this.lineasControl[J]) {
-                this.programCounter.setValor((byte) (this.bus.getValor() & 0b1111));
+                this.programCounter.setValor((int) (this.bus.getValor() & 0b1111));
                 this.notificarCambioPC();
                 EventLog.getEventLog().addEntrada("Program Counter cambia por la bandera J");
             }
@@ -646,7 +646,7 @@ public class SistemaSAP implements IClockObserver {
         return this.lineasControl;
     }
 
-    public byte getStepCount() {
+    public int getStepCount() {
         return this.stepCount;
     }
 
@@ -660,7 +660,7 @@ public class SistemaSAP implements IClockObserver {
 
     // Método auxiliar para decodificar el contenido de un registro
     public String decodificarRegistro(TipoRegistro t, int bitPos) {
-        byte val = 0;
+        int val = 0;
         if (t != null) 
             switch (t) {
                 case A:
